# Backend/src/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API - funLearn",
    "description": "Documentação contendo detalhes de como utilizar e como configurar as APIs a serem utilizadas pelo front-end."
  },
  "host": "localhost:3333",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "paths": {
...
    "/administrator/auth/session": {
      "post": {
        "tags": [
          "Administrator"
        ],
        "description": "Endpoint para autenticar o usuário administrador.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Ok."
          },
          "401": {
            "description": "Não autorizado."
          },
          "404": {
            "description": "Administrador não encontrado. \nToken não encontrado."
          }
        }
      }
    },
...
    "/teacher/auth/session": {
      "post": {
        "tags": [
          "Teacher"
        ],
        "description": "Endpoint para autenticar o usuário professor.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Ok."
          },
          "401": {
            "description": "Não autorizado."
          },
          "404": {
            "description": "Professor não encontrado. \nToken não encontrado."
          }
        }
      }
    },
...
    "/student/auth/session": {
      "post": {
        "tags": [
          "Student"
        ],
        "description": "Endpoint para autenticar o usuário estudante.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "404": {
            "description": "Estudante não encontrado. \nToken não encontrado."
          }
        }
      }
    },
...
    "/student/many": {
      "post": {
        "tags": [
          "Student"
        ],
        "description": "Endpoint para cadastrar usuários Estudentes através da importação de arquivo.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "name, teamID",
            "in": "body",
            "description": " name: Deverá ser preenchido com o nome do usuário; \n   teamID: Deverá ser preenchido com o identificador da turma.",
            "requerid": true,
            "schema": {
              "$ref": "#/definitions/studentFile"
            }
          },
          {
            "name": "file",
            "in": "file",
            "description": " file: Arquivo que deverá ser importado seguindo o modelo disponibilizado para importar os alunos.",
            "requerid": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuários cadastrados com sucesso",
            "schema": {
              "$ref": "#/definitions/CreateManystudent200"
            }
          },
          "401": {
            "description": "Não autorizado."
          },
          "404": {
            "description": "Turma não encontrada; \nTurma Inativa; \n Arquivo não importado. \n Turma ou Senha não enviada na requisição "
          },
          "500": {
            "description": "Relacionado a erro de persistência no banco de dados. \n **É necessário verificar o atributo CODE para verificar o status da requisição",
            "schema": {
              "$ref": "#/definitions/CreateManystudents500"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "studentFile": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string",
          "example": "********"
        },
        "teamID": {
          "type": "string",
          "example": "4531b848-15c8-4dfe-baf8-564491fd4a04"
        }
      }
    },
...
    "CreateManystudents500": {
      "type": "object",
      "properties": {
        "header": {
          "type": "string",
          "example": "Error"
        },
        "code": {
          "type": "number",
          "example": 500
        },
        "description": {
          "type": "object",
          "properties": {
            "code": {
              "type": "string",
              "example": "P2002"
            },
            "clientVersion": {
              "type": "string",
              "example": "4.3.1"
            },
            "meta": {
              "type": "object",
              "properties": {
                "target": {
                  "type": "string",
                  "example": "student_register_key"
                }
              }
            }
          }
        }
      }
    },
    "CreateManystudent200": {
      "type": "object",
      "properties": {
        "header": {
          "type": "string",
          "example": "Success"
        },
        "code": {
          "type": "number",
          "example": 200
        },
        "description": {
          "type": "object",
          "properties": {
            "msg": {
              "type": "string",
              "example": "Cadastro efetuado com sucesso."
            },
            "count": {
              "type": "number",
              "example": 3
            }
          }
        }
      }
    },
...
  }
}